Refactor app entry and routes files
