+++ packages/web-app/src/index.tsx
@@ -1,9 +1,28 @@
+import './ui/index.css';
+
 import React from 'react';
 import ReactDOM from 'react-dom';
+
 import Logger from 'js-logger';
-import App from './app';
+import Auth from '@aws-amplify/auth';
+import { Provider } from 'react-redux';
+import ApolloClient from 'apollo-client';
+import { ApolloProvider } from 'react-apollo';
+import { createHttpLink } from 'apollo-link-http';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { ThemeProvider } from '@material-ui/core/styles';
+
+import theme from './ui/theme';
+import initStore from './config/store';
+import awsConfig from './config/aws-config';
 import reportWebVitals from './reportWebVitals';
-import './index.css';
+import * as ApiFetch from './utils/aws-api-fetch';
+
+import Routes from './root-routes';
+
+Auth.configure(awsConfig);
+ApiFetch.configure(awsConfig);
+const store = initStore();
 
 Logger.setLevel(Logger.INFO);
 // eslint-disable-next-line react-hooks/rules-of-hooks
@@ -11,9 +30,22 @@ Logger.useDefaults();
 
 console.log('ENV', window.env);
 
+const client = new ApolloClient({
+    link: createHttpLink({
+        fetch: ApiFetch.awsApiFetch,
+    }),
+    cache: new InMemoryCache(),
+});
+
 ReactDOM.render(
     <React.StrictMode>
-        <App />
+        <ThemeProvider theme={theme}>
+            <Provider store={store}>
+                <ApolloProvider client={client}>
+                    <Routes />
+                </ApolloProvider>
+            </Provider>
+        </ThemeProvider>
     </React.StrictMode>,
     document.getElementById('root')
 );
