--- packages/web-app/src/routes/index.tsx
@@ -1,3 +1,0 @@
-import Routes from './routes';
-
-export default Routes;
