+++ packages/web-app/src/screens/profile-screen/profile-screen.test.tsx
@@ -7,7 +7,7 @@ import ProfileScreen from './profile-screen';
 describe('Profile Screen', () => {
     it('should show login form', async () => {
         renderWithAllProviders(<ProfileScreen />, { initialState: { auth: { isLoading: false, error: null, user: null } } });
-
+        // test commit
         expect(await screen.findByText(/Sign in with Email/i)).toBeInTheDocument();
     });
 });
