+++ packages/web-app/src/ui/index.css
@@ -1,7 +1,7 @@
 body {
-  height: 100%;
-  width: 100%;
-  margin: 0;
+    height: 100%;
+    width: 100%;
+    margin: 0;
 }
 
 /* body,
